#39 created abstract class for steering that joystick and carControl can extend
